FirstCommit

+++ Server/src/routes/memories.ts
@@ -1,3 +1,4 @@
+import { request } from 'http';
 import { FastifyInstance } from "fastify";
 import {z} from 'zod'
 import { prisma } from "../lib/prisma";
@@ -5,8 +6,17 @@ import { prisma } from "../lib/prisma";
 
 
 export async function memoriesRoutes(app: FastifyInstance){
-    app.get('/memories', async () => {
+
+    app.addHook('preHandler', async (request) =>{
+        await request.jwtVerify()
+    })
+    app.get('/memories', async (request) => {
+   
+        
         const memories = await prisma.memory.findMany({
+            where: {
+                userId: request.user.sub
+            },
             orderBy:{
                 createdAt: 'asc',
             }
@@ -20,7 +30,7 @@ export async function memoriesRoutes(app: FastifyInstance){
         })
     })
 
-    app.get('/memories/:id', async (request) => {
+    app.get('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid()
         })
@@ -32,6 +42,10 @@ export async function memoriesRoutes(app: FastifyInstance){
                 id
             }
         })
+
+        if(!memory.isPublic && memory.userId !== request.user.sub){
+            return reply.status(404).send()
+        }
         return memory
     })
 
@@ -48,13 +62,13 @@ export async function memoriesRoutes(app: FastifyInstance){
                 content,
                 coverUrl,
                 isPublic,
-                userId: 'fc72ccac-4f16-44e6-895e-17d0d6fe5718'
+                userId: request.user.sub
             }
         })
         return memory
     })
 
-    app.put('/memories/:id', async (request) => {
+    app.put('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid()
         })
@@ -69,7 +83,17 @@ export async function memoriesRoutes(app: FastifyInstance){
         })
         const { content, coverUrl, isPublic } = bodySchema.parse(request.body) 
 
-        const memory =  await prisma.memory.update({
+        let memory = await prisma.memory.findFirstOrThrow({
+            where:{
+                id
+            }
+        })
+
+        if( memory.userId !== request.user.sub){
+            return reply.status(404).send()
+        }
+
+       memory =  await prisma.memory.update({
             where:{
                 id
             },
@@ -82,12 +106,22 @@ export async function memoriesRoutes(app: FastifyInstance){
         return memory        
     })
 
-    app.delete('/memories/:id', async (request) => {
+    app.delete('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid()
         })
     
         const { id } = paramsSchema.parse(request.params)
+
+        const memory = await prisma.memory.findFirstOrThrow({
+            where:{
+                id
+            }
+        })
+
+        if( memory.userId !== request.user.sub){
+            return reply.status(404).send()
+        }
 
         await prisma.memory.delete({
             where:{
